Added a check for inactive channels

+++ src/commands/stuckchannel.js
@@ -40,18 +40,23 @@ module.exports = class Stuckchannel extends Command
 
         const stuckChannels = guild.channels.cache.array().filter(channel => channel.name.includes(channelPrefix));
         stuckChannels.forEach(channel => {
-            log.info(`Deleting #${channel.name} (${channel.id})`);
-            transientChannels.destroyChannel(channel.id);
-            channel.delete();
+            transientChannels.inactiveChannel(channel.id).then(inactiveChannels => {
+                // We only destroy channels that do not have a ban/suspended user in there
+                if (inactiveChannels.length === 0) {
+                    log.info(`Deleting #${channel.name} (${channel.id})`);
+
+                    transientChannels.destroyChannel(channel.id);
+                    channel.delete();
+                }
+            });
         })
 
+        log.info(`Successfully removed any stuck channels!`);
         botClient.auditLog({
             embed: new MessageEmbed()
                 .setTitle('Stuck Channel')
                 .setDescription(`Channels with prefix \`${channelPrefix}\` were removed!`)
                 .addField("Moderator", message.author, false)
         });
-
-        log.info(`Successfully removed any stuck channels!`);
     }
 };
